LayoutTests/http/tests/security/frameNavigation/navigation-DENIED-popup-child-form-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to initiate a navigation change for frame with URL http://localhost:8000/security/frameNavigation/resources/ready.html from frame with URL http://127.0.0.1:8000/security/frameNavigation/navigation-DENIED-popup-child-form.html.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to initiate a navigation change for frame with URL http://localhost:8000/security/frameNavigation/resources/ready.html from frame with URL http://127.0.0.1:8000/security/frameNavigation/navigation-DENIED-popup-child-form.html.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/frameNavigation/resources/ready.html from frame with URL http://127.0.0.1:8000/security/frameNavigation/navigation-DENIED-popup-child-form.html. Domains, protocols and ports must match.

Navigating frame.
Navigation complete.
Window count: 3
popup2: false
Done.
